robot_sdk/python/pyproject.toml
[project]
name = "robomesh-sdk"
version = "0.1.0"
description = "Python client library for robots communicating with Robomesh Roboserver"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=41.0",
]

[project.optional-dependencies]
mqtt = [
    "paho-mqtt>=2.0",
]
all = [
    "paho-mqtt>=2.0",
]
dev = [
    "pytest>=7.0",
    "paho-mqtt>=2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
